// apps/console/src/models/auth/loginSchema.ts
import { z } from "zod";

// Define Zod schema for validation
export const loginSchema = z.object({
  email: z.string().email({ message: "Invalid email address" }),
  password: z.string().min(6, { message: "Password must be at least 6 characters long" }),
  remember: z.boolean().optional(),
});